FIX: Rework généricité

+++ exam/src/main/java/com/projectexam/exam/Repositories/ConsultationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.projectexam.exam.Models.Consultation;
 
-public interface ConsultationRepository extends JpaRepository<Consultation, Long> {
+public interface ConsultationRepository extends JpaRepository<Consultation, Integer> {
     
 }
