+++ exam/src/main/java/com/projectexam/exam/Repositories/MedicamentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.projectexam.exam.Models.Medicament;
 
-public interface MedicamentRepository extends JpaRepository<Medicament, Long> {
+public interface MedicamentRepository extends JpaRepository<Medicament, Integer> {
     
 }
